Update tests for floating point inaccuracies
Due to floating point differences in i386 versus x86-64, tests were
failing on i386.

Fixes GitHub issue #178.  As indicated in comments there, the long-term
plan is to use strtod everywhere, but this fixes the test suite for
purposes such as Ubuntu packaging that need to build and test the
current version.