hooks: fix exception handling for Python functions called from C code
problem: basically the NULL pointer dereference
When an exception happens in a Python function, you get NULL from
PyObject_CallFunctionObjArgs() (instead of a valid PyObject pointer).
The value then is dereferenced and you get in trouble.

So this commit adds the NULL check with corresponding unit tests.

Also it adds few minor fixups:
 - Don't execute internal ujson's API callbacks (callPreEncodeHook() and
   callObjectHook()) when these hooks are not actually passed from the
   Python code. This is a very minor optimization, saves one C function
   call per object.
 - Get rid of the GET_EP() macro. It is used only once and therefore
   is redundant.