hooks: add pre_encode_primitive flag for the ujson.encode() function
The pre_encode_hook() is called for every single JSON value.
This is slow, especially if you have a lot of objects and complicated
hook function.

This performance overhead may be reduced if you don't call the
pre_encode_hook() for primitive types like Number and String
(since usually you don't need to handle these types).

So this commits makes this behavior default. By default you get
good performance, but the hook is not called for primitive types.
and this behavior can be disabled like this:
  ujson.encode(pre_encode_primitive=True)