Make ujson PEP-384 compliant
Add a function to check if an object is of type `decimal.Decimal`.
Since that type was previously cached as a static variable, this commit
makes it a member of the module state instead. Add the associated module
state machinery.

Only enable compact ASCII shortcut in non Limited API.

Also check if the module exists before creating it anew in the init
function.

Also remove unnecessary and leaky Py_INCREF. PyObject_GetAttrString
returns a new reference.

See PEP 384 (Defining a Stable ABI):
https://www.python.org/dev/peps/pep-0384/ and PEP 3121 (Extension Module
Initialization and Finalization):
https://www.python.org/dev/peps/pep-3121/