Fix wrap around issues on large numerics
There seemed to be a mistake on LLONG_MAX vs ULLONG_MAX (the latter is
the maximum value for an object of type `unsigned long long int` which
JSUINT64 should be). I fixed that.

I am unsure about how much performance is lost on the check for overflow
on digit check...